pyproject.toml
[project]
name = "passuth"
version = "0.1.0"
authors = [{ name = "Dowon", email = "ks2515@naver.com" }]
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
]
keywords = ["password", "argon2"]
dependencies = []

[project.optional-dependencies]
test = ["pytest", "pytest-pretty", "hypothesis", "argon2-cffi"]

[project.urls]
source = "https://github.com/Bing-su/passuth"

[project.scripts]
passuth = "passuth.__main__:main"

[dependency-groups]
dev = ["ruff", "pre-commit-uv"]

[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module"]
profile = "release"

[tool.uv]
default-groups = "all"

[tool.ruff.lint]
select = [
    "A",
    "B",
    "C4",
    "C90",
    "E",
    "EM",
    "F",
    "FA",
    "FBT",
    "FURB",
    "I001",
    "ISC",
    "N",
    "PD",
    "PERF",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TC",
    "TRY",
    "UP",
    "W",
]
ignore = ["B905", "E501"]
unfixable = ["F401"]

[tool.ruff.lint.per-file-ignores]
"tests/*_test.py" = ["S101"]
